\name{RcppMovStat-package}
\alias{RcppMovStat-package}
\alias{RcppMovStat}
\docType{package}
\title{
  Fast Moing Statitics Caculation via Rcpp
}
\description{
  This is a package providing several efficient functions to 
  calculate common moving(or rolling, running) statitics: moving average, moving median, and 
  moving maximum(minimum) . Built on C++, these functions are apparently more efficient than 
  those written in traditional R.
}
\details{
  See the following instructions for further details.
}
\author{
Pengfei (Leonard) Li, pli3@tulane.edu.

Maintainer: Pengfei (Leonard) Li <pli3@tulane.edu>
}
\keyword{R, Rcpp, Moving Average, Moving(Sliding) Window}

